requirements.txt
fastapi
uvicorn
gradio
torch
transformers
sentence-transformers
faiss-cpu
accelerate
bitsandbytes
